% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{apply_grade_weighting}
\alias{apply_grade_weighting}
\title{Apply GRADE-based Weighting}
\usage{
apply_grade_weighting(weights, grade_scores)
}
\arguments{
\item{weights}{Base weights to modify}

\item{grade_scores}{GRADE quality scores}
}
\value{
Modified weights incorporating GRADE scores
}
\description{
Apply GRADE-based Weighting
}
