% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cbamm_config}
\alias{cbamm_config}
\title{Package configuration settings..}
\usage{
cbamm_config(
  methods = list(transport = TRUE, hksj = TRUE, bayesian = TRUE, robust_variance = TRUE,
    pet_peese = TRUE, conflict_detection = TRUE, missing_studies = TRUE, grade_weighting
    = FALSE),
  estimators = c("REML", "DL", "PM", "ML"),
  output = list(export = FALSE, interactive = FALSE, verbose = TRUE, plots = TRUE),
  transportability = list(method = "entropy", target_population = NULL, covariates =
    NULL),
  bayesian = list(method = "brms", chains = 4, iter = 2000, warmup = 1000, cores = 1,
    prior = "weakly_informative"),
  sensitivity = list(egger_test = TRUE, begg_test = TRUE, trim_fill = TRUE,
    copas_selection = FALSE, publication_bias_methods = c("egger", "begg", "trim_fill"))
)
}
\arguments{
\item{methods}{List of methods to enable/disable}

\item{estimators}{Vector of variance estimators to try}

\item{output}{List of output options}

\item{transportability}{List of transportability settings}

\item{bayesian}{List of Bayesian method settings}

\item{sensitivity}{List of sensitivity analysis settings}
}
\value{
List containing configuration options
}
\description{
Creates configuration object for CBAMM analysis
}
\examples{
# Default configuration
config <- cbamm_config()

# Custom configuration
config <- cbamm_config(
  methods = list(transport = FALSE, hksj = TRUE),
  estimators = c("REML", "DL")
)
}
