% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_package}
\alias{check_package}
\title{Utility Functions for CBAMM Package.}
\usage{
check_package(pkg, quietly = TRUE)
}
\arguments{
\item{pkg}{Package name}

\item{quietly}{Whether to suppress messages}
}
\value{
Logical indicating if package is available
}
\description{
Internal utility functions used throughout CBAMM
Check if Package is Available.
}
\details{
Safely check if a package is available and can be loaded
}
\keyword{internal}
