% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{compute_bayesian_stacking}
\alias{compute_bayesian_stacking}
\title{Compute Bayesian Model Stacking}
\usage{
compute_bayesian_stacking(bayesian_results, data, config)
}
\arguments{
\item{bayesian_results}{List of Bayesian results from different methods}

\item{data}{Original data}

\item{config}{Configuration}
}
\value{
Stacking results
}
\description{
Compute Bayesian Model Stacking
}
\keyword{internal}
