% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_robust_variance}
\alias{compute_robust_variance}
\title{Robust Variance Calculation.}
\usage{
compute_robust_variance(data, cluster_var = NULL, method = "CR2")
}
\arguments{
\item{data}{Data frame}

\item{cluster_var}{Clustering variable}

\item{method}{Robust variance method}
}
\value{
Robust variance estimates
}
\description{
Calculate robust variance estimates for clustered data
}
\keyword{internal}
