% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_effect_size}
\alias{convert_effect_size}
\title{Transform effect size metrics. Metrics.}
\usage{
convert_effect_size(yi, from = "Cohen_d", to = "Hedges_g", ...)
}
\arguments{
\item{yi}{Effect sizes}

\item{from}{Source metric}

\item{to}{Target metric}

\item{...}{Additional parameters for conversion}
}
\value{
Converted effect sizes
}
\description{
Convert between different effect size metrics
}
