% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Generate summary tables..}
\usage{
create_summary_table(results_list, round_digits = 3)
}
\arguments{
\item{results_list}{List of results to summarize}

\item{round_digits}{Number of digits to round to}
}
\value{
Data frame with formatted results
}
\description{
Create a formatted summary table for results
}
