% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{.compute_weightit_weights}
\alias{.compute_weightit_weights}
\title{Compute pure R propensity score weighting-based Transport Weights}
\usage{
.compute_weightit_weights(X, target_population, truncation)
}
\description{
Compute pure R propensity score weighting-based Transport Weights
}
\keyword{internal}
