% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{extract_fixed_effects}
\alias{extract_fixed_effects}
\title{Extract Fixed Effects}
\usage{
extract_fixed_effects(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Fixed effects estimates
}
\description{
Extract fixed effects from pure R Bayesian model
}
