% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{extract_random_effects}
\alias{extract_random_effects}
\title{Extract Random Effects}
\usage{
extract_random_effects(model)
}
\arguments{
\item{model}{Model object}
}
\value{
Random effects estimates
}
\description{
Extract random effects from pure R Bayesian model
}
