% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{fast_robust_rma}
\alias{fast_robust_rma}
\title{Fast Robust Meta-Analysis Wrapper}
\usage{
fast_robust_rma(
  yi,
  sei,
  data = NULL,
  method = "REML",
  weights = NULL,
  mods = NULL,
  use_hksj = TRUE
)
}
\arguments{
\item{yi}{Effect sizes}

\item{sei}{Standard errors}

\item{data}{Optional data frame}

\item{method}{Variance estimator method}

\item{weights}{Optional study weights}

\item{mods}{Optional moderators}

\item{use_hksj}{Whether to use Hartung-Knapp-Sidik-Jonkman adjustment}
}
\value{
Meta-analysis results object
}
\description{
Fast Robust Meta-Analysis Wrapper
}
\keyword{internal}
