% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{generate_cbamm_plots}
\alias{generate_cbamm_plots}
\title{Generate CBAMM Plots}
\usage{
generate_cbamm_plots(results, config)
}
\arguments{
\item{results}{CBAMM results object}

\item{config}{CBAMM configuration}
}
\value{
List of plots
}
\description{
Generate CBAMM Plots
}
\keyword{internal}
