% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{get_bayesian_priors}
\alias{get_bayesian_priors}
\title{Get Bayesian Priors}
\usage{
get_bayesian_priors(prior_type, method = "brms")
}
\arguments{
\item{prior_type}{Type of prior ("weakly_informative", "informative", "non_informative")}

\item{method}{Either "brms" or "pure R MCMC approximation"}
}
\value{
Prior specifications
}
\description{
Get Bayesian Priors
}
\keyword{internal}
