% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{get_jags_model_string}
\alias{get_jags_model_string}
\title{Get pure R MCMC approximation Model String}
\usage{
get_jags_model_string(prior_type = "weakly_informative")
}
\arguments{
\item{prior_type}{Type of prior}
}
\value{
pure R MCMC approximation model specification as string
}
\description{
Get pure R MCMC approximation Model String
}
\keyword{internal}
