% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R,
%   R/sensitivity_backup_20250904_120326.R,
%   R/sensitivity_backup_20250904_124954.R
\name{pet_peese}
\alias{pet_peese}
\title{PET-PEESE Analysis}
\usage{
pet_peese(yi, sei)

pet_peese(yi, sei)

pet_peese(yi, sei)
}
\arguments{
\item{yi}{Effect sizes}

\item{sei}{Standard errors}
}
\value{
Vector with PET and PEESE intercepts

Vector with PET and PEESE intercepts

Vector with PET and PEESE intercepts
}
\description{
PET-PEESE Analysis

PET-PEESE Analysis

PET-PEESE Analysis
}
