% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{pure_r_bayesian_meta}
\alias{pure_r_bayesian_meta}
\title{Pure R Bayesian Meta-Analysis}
\usage{
pure_r_bayesian_meta(formula, data, family = "gaussian", ...)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Input data}

\item{family}{Model family}

\item{...}{Additional arguments}
}
\value{
Bayesian model results
}
\description{
Provides basic Bayesian meta-analysis using pure R implementation
}
