% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{pure_r_coef_test}
\alias{pure_r_coef_test}
\title{Pure R Coefficient Test}
\usage{
pure_r_coef_test(model, vcov)
}
\arguments{
\item{model}{Model object}

\item{vcov}{Variance-covariance matrix}
}
\value{
Test results
}
\description{
Perform coefficient tests using pure R implementation
}
