% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{pure_r_mcmc_chain}
\alias{pure_r_mcmc_chain}
\title{Pure R MCMC Chain}
\usage{
pure_r_mcmc_chain(samples)
}
\arguments{
\item{samples}{MCMC samples}
}
\value{
MCMC chain object
}
\description{
Convert MCMC results to chain format
}
