% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{pure_r_mcmc_model}
\alias{pure_r_mcmc_model}
\title{Pure R MCMC Model}
\usage{
pure_r_mcmc_model(model_string, data, n_chains = 3)
}
\arguments{
\item{model_string}{Model specification}

\item{data}{Input data}

\item{n_chains}{Number of chains}
}
\value{
MCMC model object
}
\description{
Basic MCMC implementation using pure R
}
