% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{pure_r_propensity_weights}
\alias{pure_r_propensity_weights}
\title{Pure R Propensity Score Weights}
\usage{
pure_r_propensity_weights(formula, data, method = "ps")
}
\arguments{
\item{formula}{Propensity score formula}

\item{data}{Input data}

\item{method}{Weighting method}
}
\value{
Propensity score weights
}
\description{
Calculate propensity score weights using pure R
}
