% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pure_r_implementations.R
\name{pure_r_robust_vcov}
\alias{pure_r_robust_vcov}
\title{Pure R Robust Variance-Covariance Matrix}
\usage{
pure_r_robust_vcov(model, cluster, type = "CR2")
}
\arguments{
\item{model}{Model object}

\item{cluster}{Clustering variable}

\item{type}{Variance estimator type}
}
\value{
Robust variance-covariance matrix
}
\description{
Compute robust variance-covariance matrix using pure R
}
