% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{run_brms_analysis}
\alias{run_brms_analysis}
\title{Run pure R Bayesian approximation Bayesian Analysis}
\usage{
run_brms_analysis(data, config)
}
\arguments{
\item{data}{Study data}

\item{config}{CBAMM configuration}
}
\value{
pure R Bayesian approximation results or NULL if failed
}
\description{
Run pure R Bayesian approximation Bayesian Analysis
}
\keyword{internal}
