% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{run_core_meta_analysis}
\alias{run_core_meta_analysis}
\title{Run Core Meta-Analysis}
\usage{
run_core_meta_analysis(data, config)
}
\arguments{
\item{data}{Study data}

\item{config}{CBAMM configuration}
}
\value{
List with meta-analysis results
}
\description{
Run Core Meta-Analysis
}
\keyword{internal}
