% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{run_jags_analysis}
\alias{run_jags_analysis}
\title{Run pure R MCMC approximation Bayesian Analysis}
\usage{
run_jags_analysis(data, config)
}
\arguments{
\item{data}{Study data}

\item{config}{CBAMM configuration}
}
\value{
pure R MCMC approximation results or NULL if failed
}
\description{
Run pure R MCMC approximation Bayesian Analysis
}
\keyword{internal}
