% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R,
%   R/sensitivity_backup_20250904_120326.R,
%   R/sensitivity_backup_20250904_124954.R
\name{run_publication_bias_sensitivity}
\alias{run_publication_bias_sensitivity}
\title{Run Publication Bias Sensitivity Analysis}
\usage{
run_publication_bias_sensitivity(data)

run_publication_bias_sensitivity(data)

run_publication_bias_sensitivity(data)
}
\arguments{
\item{data}{Study data}
}
\value{
List of publication bias test results

List of publication bias test results

List of publication bias test results
}
\description{
Run Publication Bias Sensitivity Analysis

Run Publication Bias Sensitivity Analysis

Run Publication Bias Sensitivity Analysis
}
\keyword{internal}
