% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R,
%   R/sensitivity_backup_20250904_120326.R,
%   R/sensitivity_backup_20250904_124954.R
\name{run_sensitivity_analyses}
\alias{run_sensitivity_analyses}
\title{Run Sensitivity Analyses}
\usage{
run_sensitivity_analyses(data, config)

run_sensitivity_analyses(data, config)

run_sensitivity_analyses(data, config)
}
\arguments{
\item{data}{Study data}

\item{config}{CBAMM configuration}
}
\value{
List of sensitivity analysis results

List of sensitivity analysis results

List of sensitivity analysis results
}
\description{
Run Sensitivity Analyses

Run Sensitivity Analyses

Run Sensitivity Analyses
}
\keyword{internal}
