% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{safe_robust_rma}
\alias{safe_robust_rma}
\title{Safe Meta-Analysis Wrapper}
\usage{
safe_robust_rma(
  yi,
  sei,
  data = NULL,
  method = "REML",
  weights = NULL,
  mods = NULL,
  use_hksj = TRUE
)
}
\description{
Safe Meta-Analysis Wrapper
}
\keyword{internal}
