% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{se_to_var}
\alias{se_to_var}
\title{Compute Variance from Standard Error}
\usage{
se_to_var(se)
}
\arguments{
\item{se}{Standard errors}
}
\value{
Variances
}
\description{
Compute Variance from Standard Error
}
