% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm.R, R/cbamm_original_20250904_134106.R
\name{summarize_input_data}
\alias{summarize_input_data}
\title{Summarize Input Data.}
\usage{
summarize_input_data(data)

summarize_input_data(data)
}
\arguments{
\item{data}{Input data frame}
}
\value{
List with data summary

List with data summary
}
\description{
Creates summary of input data for results object

Creates summary of input data for results object
}
\keyword{internal}
