% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbamm_original_20250904_134106.R, R/methods.R
\name{summary.cbamm_results}
\alias{summary.cbamm_results}
\title{Summary method for cbamm_results}
\usage{
\method{summary}{cbamm_results}(object, ...)

\method{summary}{cbamm_results}(object, ...)
}
\arguments{
\item{object}{CBAMM results object}

\item{...}{Additional arguments (unused)}
}
\value{
Summary data frame
}
\description{
Summary method for cbamm_results

Summary Method for CBAMM Results
}
