% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{update_config}
\alias{update_config}
\title{Update Package configuration settings..}
\usage{
update_config(config, ...)
}
\arguments{
\item{config}{Existing cbamm_config object}

\item{...}{Named arguments to update}
}
\value{
Updated configuration object
}
\description{
Update an existing configuration object
}
